refactor: reorganize the structure of the folders of api
