+++ web/src/api/tasks/fetch-tasks/query-options.ts
@@ -1,0 +1,6 @@
+import { fetchTasks, FetchTasksParams } from "./queries";
+
+export const tasksQueryOptions = ({ status }: FetchTasksParams) => ({
+  queryKey: ["tasks", status],
+  queryFn: () => fetchTasks({ status }),
+});
